refactor(credential)!: collapse CredentialDispatch into the capability bitflag (ADR-0088 D3)
The runtime CredentialDispatch table held three bool flags (refreshable /
testable / revocable) that mirrored sub-trait membership the CredentialRegistry
already computes into its Capabilities bitflag at register time. Per ADR-0088
D3 ("capability = closure/strategy presence, not a parallel flag table"),
delete that parallel table and read capability from the one source of truth.

- Add `CredentialRegistry::is_refreshable/is_testable/is_revocable(key)` —
  thin reads over the computed Capabilities bitflag.
- Repoint every `CredentialService` capability/contains check from
  `self.dispatch.*` to `self.registry.*` (resolve/test/refresh/revoke guards +
  type_info). Behavior-preserving: the bitflag is computed from the same
  sub-traits the dispatch flags mirrored (and is correct even where a
  `mark_*` call was previously omitted).
- Delete `dispatch.rs` (CredentialDispatch) and the test-only
  `registration.rs` registrar (`register_credential_complete` had no
  production caller — registration goes straight through
  `CredentialRegistry::register`). Drop `dispatch` from
  `CredentialServiceBuilder::new` / `CredentialService::__from_parts` and the
  test fixtures.
- Relocate `DispatchError` (the operation-dispatch table's own registration
  error) from the deleted dispatch.rs into ops.rs; `DispatchOps<B,PS>` (the
  real async-op closures) is unchanged.
- Delete `registry_sync_invariant` + `composite_registration` tests (they
  cross-checked drift between the now-single table; dup-KEY fail-closed stays
  covered by `runtime_duplicate_key_fatal`).

N7 supply-chain defense is unchanged — `CredentialRegistry` dup-KEY remains
fatal (§15.6). No production code wired `CredentialServiceBuilder` yet, so the
builder API change has no external callers.
