+++ crates/credential-runtime/src/builder.rs
@@ -23,7 +23,6 @@ use nebula_storage::credential::{
 };
 use tokio_util::sync::CancellationToken;
 
-use crate::dispatch::CredentialDispatch;
 use crate::observer::CredentialObserver;
 use crate::ops::DispatchOps;
 use crate::service::CredentialService;
@@ -40,7 +39,6 @@ pub struct CredentialServiceBuilder<B: CredentialStore, PS: PendingStateStore> {
     cache_config: CacheConfig,
     pending_store: PS,
     registry: Arc<CredentialRegistry>,
-    dispatch: Arc<CredentialDispatch>,
     ops: Arc<DispatchOps<B, PS>>,
     observer: Arc<dyn CredentialObserver>,
     lease_config: LeaseLifecycleConfig,
@@ -60,7 +58,6 @@ impl<B: CredentialStore, PS: PendingStateStore> CredentialServiceBuilder<B, PS> 
         cache_config: CacheConfig,
         pending_store: PS,
         registry: Arc<CredentialRegistry>,
-        dispatch: Arc<CredentialDispatch>,
         ops: Arc<DispatchOps<B, PS>>,
         observer: Arc<dyn CredentialObserver>,
         lease_config: LeaseLifecycleConfig,
@@ -73,7 +70,6 @@ impl<B: CredentialStore, PS: PendingStateStore> CredentialServiceBuilder<B, PS> 
             cache_config,
             pending_store,
             registry,
-            dispatch,
             ops,
             observer,
             lease_config,
@@ -145,7 +141,6 @@ impl<B: CredentialStore, PS: PendingStateStore> CredentialServiceBuilder<B, PS> 
             lease,
             self.pending_store,
             self.registry,
-            self.dispatch,
             self.ops,
             self.observer,
             self.external,
