+++ crates/credential-runtime/src/lib.rs
@@ -16,11 +16,9 @@
 
 pub mod binding;
 pub mod builder;
-pub mod dispatch;
 pub mod error;
 pub mod observer;
 pub mod ops;
-pub mod registration;
 pub mod scope;
 pub mod service;
 pub mod state_source;
@@ -33,14 +31,12 @@ pub mod test_fixtures;
 
 pub use binding::{TenantFingerprint, ValidatedCredentialBinding, ValidatedCredentialBindingError};
 pub use builder::CredentialServiceBuilder;
-pub use dispatch::{CredentialDispatch, DispatchError};
 pub use error::CredentialServiceError;
 pub use observer::{CredentialObserver, EventMetricObserver, NoopObserver};
 pub use ops::{
-    DispatchOps, register_all_builtin_ops, register_interactive_ops, register_refreshable_ops,
-    register_revocable_ops, register_runtime_ops, register_testable_ops,
+    DispatchError, DispatchOps, register_all_builtin_ops, register_interactive_ops,
+    register_refreshable_ops, register_revocable_ops, register_runtime_ops, register_testable_ops,
 };
-pub use registration::{RegistrationError, register_credential_complete};
 pub use scope::{FixedScopeResolver, TenantScope};
 pub use service::{
     Acquisition, CredentialService, CredentialTypeInfo, LayeredStore, TestReport, TypeCapabilities,
